fix(visualization): move view type dropdown component to shared folder
